+++ stories/navbar.stories.tsx
@@ -20,7 +20,7 @@ storiesOf('Navbar', module)
     <Navbar
       navItems={[
         {
-          type: 'icon',
+          type: 'image',
           src:
             'https://raw.githubusercontent.com/HospitalRun/hospitalrun.github.io/master/favicon.png',
           onClick: () => {
@@ -67,7 +67,7 @@ storiesOf('Navbar', module)
       variant="dark"
       navItems={[
         {
-          type: 'icon',
+          type: 'image',
           src:
             'https://raw.githubusercontent.com/HospitalRun/hospitalrun.github.io/master/favicon.png',
           onClick: () => {
@@ -131,7 +131,7 @@ storiesOf('Navbar', module)
     <Navbar
       navItems={[
         {
-          type: 'icon',
+          type: 'image',
           src:
             'https://raw.githubusercontent.com/HospitalRun/hospitalrun.github.io/master/favicon.png',
           onClick: () => {
@@ -176,13 +176,10 @@ storiesOf('Navbar', module)
         {
           type: 'link-list-icon',
           alignRight: true,
-          label: 'User',
-          src:
-            'https://s3.us-east-2.amazonaws.com/upload-icon/uploads/icons/png/7104608081548233620-128.png',
-          onClick: () => {
-            Toast('success', 'Settings icon clicked!!', 'Success')
-          },
-          className: 'ml-3',
+          label: 'Patient',
+          name: 'patient',
+          size: 'lg',
+          iconClassName: 'align-bottom',
           children: [
             {
               type: 'link',
@@ -202,20 +199,24 @@ storiesOf('Navbar', module)
         },
         {
           type: 'icon',
-          src: 'https://image.flaticon.com/icons/svg/126/126472.svg',
+          label: 'Patient',
+          name: 'patient',
           onClick: () => {
             Toast('success', 'Settings icon clicked!!', 'Success')
           },
-          className: 'ml-4',
+          size: 'lg',
+          outline: true,
+          iconClassName: 'align-bottom',
         },
         {
           type: 'icon',
-          src:
-            'https://s3.us-east-2.amazonaws.com/upload-icon/uploads/icons/png/7104608081548233620-128.png',
+          label: 'Patient',
+          name: 'patient',
           onClick: () => {
-            Toast('success', 'Profile icon clicked!!', 'Success')
+            Toast('success', 'Settings icon clicked!!', 'Success')
           },
-          className: 'ml-3',
+          size: 'lg',
+          iconClassName: 'align-bottom',
         },
       ]}
     />
@@ -234,7 +235,7 @@ storiesOf('Navbar', module)
           className: 'ml-4',
         },
         {
-          type: 'icon',
+          type: 'image',
           src:
             'https://raw.githubusercontent.com/HospitalRun/hospitalrun.github.io/master/favicon.png',
           onClick: () => {
@@ -276,7 +277,7 @@ storiesOf('Navbar', module)
     <Navbar
       navItems={[
         {
-          type: 'icon',
+          type: 'image',
           src:
             'https://raw.githubusercontent.com/HospitalRun/hospitalrun.github.io/master/favicon.png',
           onClick: () => {