+++ test/navbar.test.tsx
@@ -173,14 +173,14 @@ describe('Navbar', () => {
     const header = NavbarWrapper.find(NavBarRB.Brand)
     expect(header.text().trim()).toEqual('Test')
   })
-  it('should render icon src as passed in src prop', () => {
+  it('should render image src as passed in src prop', () => {
     const onClickButton = sinon.spy()
     const onChangeInput = sinon.spy()
     const NavbarWrapper = shallow(
       <Navbar
         navItems={[
           {
-            type: 'icon',
+            type: 'image',
             label: 'test',
             src: 'anysource',
           },
@@ -200,8 +200,8 @@ describe('Navbar', () => {
         ]}
       />,
     )
-    const iconImg = NavbarWrapper.find('img')
-    expect(iconImg).toHaveLength(1)
+    const image = NavbarWrapper.find('img')
+    expect(image).toHaveLength(1)
   })
 
   it('should render button color as passed in buttonColor prop', () => {